+++ databases/migrations/2018_01_09_043202_create_users_table.py
@@ -6,23 +6,21 @@ class CreateUsersTable(Migration):
 
     def up(self):
         """Run the migrations."""
-        with self.schema.create('users') as table:
-            table.increments('id')
-            table.string('name')
-            table.string('email').unique()
-            table.string('password')
-            table.string('second_password').nullable()
-            table.string('remember_token').nullable()
-            table.timestamp('verified_at').nullable()
+        with self.schema.create("users") as table:
+            table.increments("id").primary()
+            table.string("name")
+            table.string("email").unique()
+            table.string("password")
+            table.string("second_password").nullable()
+            table.string("remember_token").nullable()
+            table.timestamp("verified_at").nullable()
             table.timestamps()
 
         if not self.schema._dry:
-            User.on(self.connection).set_schema(self.schema_name).create({
-                'name': 'Joe',
-                'email': 'joe@email.com',
-                'password': 'secret'
-            })
+            User.on(self.connection).set_schema(self.schema_name).create(
+                {"name": "Joe", "email": "joe@email.com", "password": "secret"}
+            )
 
     def down(self):
         """Revert the migrations."""
-        self.schema.drop('users')
+        self.schema.drop("users")