+++ databases/migrations/2020_04_17_000000_create_friends_table.py
@@ -1,19 +1,20 @@
 from src.masoniteorm.migrations.Migration import Migration
 
+
 class CreateFriendsTable(Migration):
 
-    def up(self):  
+    def up(self):
         """
         Run the migrations.
         """
 
-        with self.schema.create('friends') as table:
-            table.increments('id')
-            table.string('name')
-            table.integer('age')
+        with self.schema.create("friends") as table:
+            table.increments("id").primary()
+            table.string("name")
+            table.integer("age")
 
     def down(self):
         """
         Revert the migrations.
         """
-        self.schema.drop('friends')
+        self.schema.drop("friends")