+++ src/masoniteorm/schema/platforms/PostgresPlatform.py
@@ -22,7 +22,6 @@ class PostgresPlatform(Platform):
         "integer": "INTEGER",
         "big_integer": "BIGINT",
         "tiny_integer": "TINYINT",
-        "big_increments": "BIGSERIAL UNIQUE",
         "small_integer": "SMALLINT",
         "medium_integer": "MEDIUMINT",
         # Postgres database does not implement unsigned types
@@ -32,7 +31,6 @@ class PostgresPlatform(Platform):
         "tiny_integer_unsigned": "TINYINT",
         "small_integer_unsigned": "SMALLINT",
         "medium_integer_unsigned": "MEDIUMINT",
-        "increments": "SERIAL UNIQUE",
         "uuid": "UUID",
         "binary": "BYTEA",
         "boolean": "BOOLEAN",
@@ -55,8 +53,13 @@ class PostgresPlatform(Platform):
         "date": "DATE",
         "year": "YEAR",
         "datetime": "TIMESTAMPTZ",
-        "tiny_increments": "TINYINT AUTO_INCREMENT",
         "unsigned": "INT",
+        "tiny_increments": "SMALLSERIAL",
+        "increments": "SERIAL",
+        "big_increments": "BIGSERIAL",
+        "tiny_increments_primary": "SMALLSERIAL PRIMARY KEY",
+        "increments_primary": "SERIAL PRIMARY KEY",
+        "big_increments_primary": "BIGSERIAL PRIMARY KEY",
     }
 
     table_info_map = {
@@ -161,9 +164,6 @@ class PostgresPlatform(Platform):
 
             constraint = ""
             column_constraint = ""
-            if column.primary:
-                constraint = "PRIMARY KEY"
-
             if column.column_type == "enum":
                 values = ", ".join(f"'{x}'" for x in column.values)
                 column_constraint = f" CHECK({column.name} IN ({values}))"