+++ tests/mssql/schema/test_mssql_schema_builder.py
@@ -1,9 +1,9 @@
 import unittest
 
-from tests.integrations.config.database import DATABASES
 from src.masoniteorm.connections import MSSQLConnection
 from src.masoniteorm.schema import Schema
 from src.masoniteorm.schema.platforms import MSSQLPlatform
+from tests.integrations.config.database import DATABASES
 
 
 class TestMSSQLSchemaBuilder(unittest.TestCase):
@@ -26,7 +26,9 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
         self.assertEqual(len(blueprint.table.added_columns), 2)
         self.assertEqual(
             blueprint.to_sql(),
-            ["CREATE TABLE [users] ([name] VARCHAR(255) NOT NULL, [age] INT NOT NULL)"],
+            [
+                "CREATE TABLE [users] ([name] VARCHAR(255) NOT NULL, [age] INT NOT NULL)"
+            ],
         )
 
     def test_can_add_tiny_text(self):
@@ -69,7 +71,10 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
 
         self.assertEqual(
             blueprint.to_sql(),
-            ["""CREATE TABLE [users] (""" """[amount] FLOAT(19, 4) NOT NULL)"""],
+            [
+                """CREATE TABLE [users] ("""
+                """[amount] FLOAT(19, 4) NOT NULL)"""
+            ],
         )
 
     def test_can_have_unsigned_columns(self):
@@ -134,30 +139,32 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
 
     def test_can_advanced_table_creation(self):
         with self.schema.create("users") as blueprint:
-            blueprint.increments("id")
+            blueprint.increments("id").primary()
             blueprint.string("name")
             blueprint.string("email").unique()
             blueprint.string("password")
             blueprint.integer("admin").default(0)
             blueprint.string("remember_token").nullable()
             blueprint.timestamp("verified_at").nullable()
-            blueprint.timestamp("registered_at").default_raw("CURRENT_TIMESTAMP")
+            blueprint.timestamp("registered_at").default_raw(
+                "CURRENT_TIMESTAMP"
+            )
             blueprint.timestamps()
 
         self.assertEqual(len(blueprint.table.added_columns), 10)
         self.assertEqual(
             blueprint.to_sql(),
             [
-                "CREATE TABLE [users] ([id] INT IDENTITY NOT NULL, [name] VARCHAR(255) NOT NULL, [email] VARCHAR(255) NOT NULL, "
+                "CREATE TABLE [users] ([id] INT PRIMARY KEY IDENTITY NOT NULL, [name] VARCHAR(255) NOT NULL, [email] VARCHAR(255) NOT NULL, "
                 "[password] VARCHAR(255) NOT NULL, [admin] INT NOT NULL DEFAULT 0, [remember_token] VARCHAR(255) NULL, "
                 "[verified_at] DATETIME NULL, [registered_at] DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP, [created_at] DATETIME NULL DEFAULT CURRENT_TIMESTAMP, "
-                "[updated_at] DATETIME NULL DEFAULT CURRENT_TIMESTAMP, CONSTRAINT users_id_primary PRIMARY KEY (id), CONSTRAINT users_email_unique UNIQUE (email))"
+                "[updated_at] DATETIME NULL DEFAULT CURRENT_TIMESTAMP, CONSTRAINT users_email_unique UNIQUE (email))"
             ],
         )
 
     def test_can_advanced_table_creation2(self):
         with self.schema.create("users") as blueprint:
-            blueprint.increments("id")
+            blueprint.increments("id").primary()
             blueprint.enum("gender", ["male", "female"])
             blueprint.string("name")
             blueprint.string("duration")
@@ -169,9 +176,9 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
             blueprint.string("thumbnail").nullable()
             blueprint.integer("premium")
             blueprint.integer("author_id").unsigned().nullable()
-            blueprint.foreign("author_id").references("id").on("users").on_delete(
-                "CASCADE"
-            )
+            blueprint.foreign("author_id").references("id").on(
+                "users"
+            ).on_delete("CASCADE")
             blueprint.text("description")
             blueprint.timestamps()
 
@@ -180,11 +187,11 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
             blueprint.to_sql(),
             (
                 [
-                    "CREATE TABLE [users] ([id] INT IDENTITY NOT NULL, [gender] VARCHAR(255) NOT NULL CHECK([gender] IN ('male', 'female')), [name] VARCHAR(255) NOT NULL, [duration] VARCHAR(255) NOT NULL, "
+                    "CREATE TABLE [users] ([id] INT PRIMARY KEY IDENTITY NOT NULL, [gender] VARCHAR(255) NOT NULL CHECK([gender] IN ('male', 'female')), [name] VARCHAR(255) NOT NULL, [duration] VARCHAR(255) NOT NULL, "
                     "[url] VARCHAR(255) NOT NULL, [last_address] VARCHAR(255) NULL, [route_origin] VARCHAR(255) NULL, [mac_address] VARCHAR(255) NULL, [published_at] DATETIME NOT NULL, [thumbnail] VARCHAR(255) NULL, [premium] INT NOT NULL, "
                     "[author_id] INT NULL, [description] TEXT NOT NULL, [created_at] DATETIME NULL DEFAULT CURRENT_TIMESTAMP, "
                     "[updated_at] DATETIME NULL DEFAULT CURRENT_TIMESTAMP, "
-                    "CONSTRAINT users_id_primary PRIMARY KEY (id), CONSTRAINT users_author_id_foreign FOREIGN KEY ([author_id]) REFERENCES [users]([id]) ON DELETE CASCADE)"
+                    "CONSTRAINT users_author_id_foreign FOREIGN KEY ([author_id]) REFERENCES [users]([id]) ON DELETE CASCADE)"
                 ]
             ),
         )
@@ -192,9 +199,9 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
     def test_can_add_columns_with_foreign_key_constraint_name(self):
         with self.schema.create("users") as blueprint:
             blueprint.integer("profile_id")
-            blueprint.foreign("profile_id", name="profile_foreign").references("id").on(
-                "profiles"
-            )
+            blueprint.foreign("profile_id", name="profile_foreign").references(
+                "id"
+            ).on("profiles")
 
         self.assertEqual(len(blueprint.table.added_columns), 1)
         self.assertEqual(
@@ -315,7 +322,9 @@ class TestMSSQLSchemaBuilder(unittest.TestCase):
 
     def test_can_change_column_enum(self):
         with self.schema.table("users") as blueprint:
-            blueprint.enum("status", ["active", "inactive"]).default("active").change()
+            blueprint.enum("status", ["active", "inactive"]).default(
+                "active"
+            ).change()
 
         self.assertEqual(len(blueprint.table.changed_columns), 1)
         self.assertEqual(